% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_factset_financial_data.R
\name{fake_factset_financial_data}
\alias{fake_factset_financial_data}
\title{Minimal fake \code{factset_financial_data} dataset that allow overwriting values}
\usage{
fake_factset_financial_data(
  fsym_id = NULL,
  isin = NULL,
  factset_entity_id = NULL,
  adj_price = NULL,
  adj_shares_outstanding = NULL,
  issue_type = NULL,
  one_adr_eq = NULL,
  ...
)
}
\arguments{
\item{fsym_id}{A FactSet FSYM ID (character).}

\item{isin}{An ISIN (character).}

\item{factset_entity_id}{A FactSet Entity ID (character).}

\item{adj_price}{Adjusted price (numeric).}

\item{adj_shares_outstanding}{Adjusted shares outstanding (numeric).}

\item{issue_type}{Issue type (character).}

\item{one_adr_eq}{One ADR equivalent (numeric).}

\item{...}{Arguments passed on to \code{\link[tibble:tibble]{tibble::tibble()}}. These are useful in
case you want to add additional columns to the output tibble.}
}
\value{
A tibble
}
\description{
Minimal fake \code{factset_financial_data} dataset that allow overwriting values
}
\section{Fake data}{

These funtions are developer-oriented. They all call \code{\link[=tibble]{tibble()}} so
you can expect all the goodies that come with that.
\itemize{
\item \code{fake_foo} fakes a dataset with the name \code{foo}
}
}

\section{Params}{

The arguments are the column names of the datasets being faked. They all have
a default and it can be overwritten.
}

\section{Pros and cons}{

By default, these functions output a 1-row data-frame, which can be useful
for CI/CD. They also help you to avoid duplicating test code, and help
the reader of your code to focus on the one thing you want to test, instead
of burring that thing in the much longer code you need to create a fake
object from scratch. Since the default parameters are abstracted into
\code{utils.R}, it makes it easier to maintain coordination of values across a
variety of datasets.

But \verb{fake_*()} functions hide the explicit content. If the reader of your
code wants to inspect the data being tested, they need to jump to the
function definition or call them interactively.
}

\examples{
fake_factset_financial_data()

fake_factset_financial_data(isin = "US0123456789")
}
